input: add fflush
The problem is that on some systems the printf does not flush and the output gets corrupted. I confirm this on musl system. However this is only the begining of the rotten can, as for example scrolling does not work correctly in tty, and I am forced to use ^l to redraw the screen every time, I can make it do that automatically, but it will be suboptimal solution, and cause high cpu spikes.